requirements.txt
-i https://pypi.org/simple
exceptiongroup==1.2.0; python_version < '3.11'
iniconfig==2.0.0; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
pyserial==3.5
pytest==7.4.3
tomli==2.0.1; python_version < '3.11'
